refactor: 분리된 CardDetails View component를 합치도록 변경
기존에는 top, bottom으로 또 컴포넌트화 시켜서 관심사를 분리하였음. 하지만 UI만 관심을 주도록 하고 순수 함수만 props로 변경하여 똑같이 사용되고 있는데 굳이 컴포넌트화 시킬 필요가 없다고 판단하였음.
내가 기대하는 CardDetailsView 컴포넌트 속 관심사는 오로지 UI이며 여기서 더 쪼개면 의미가 없고 오히려 가독성을 해칠 수 있다고 판단하였음.
